# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: enrichment v3.0
  description: >-
  contact: {}
host: localhost:3000
basePath: /
securityDefinitions: {}
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
  /predict_needs:
    post:
      summary: predict_needs
      tags:
      - prediction
      operationId: predict_needs
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          example:
          - id: 1
            age: 64
            gender: m
            location: toscana
            education: scuola secondaria di II grado
            profession: lavoratore indipendente
          type: array
          items:
            $ref: '#/definitions/predict_needs_request'
      responses:
        '200':
          description: ''
          headers: {}
      security: []
  /predict_cultures:
    post:
      summary: predict_cultures
      tags:
      - prediction
      operationId: predict_cultures
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          example:
          - id: 1
            age: 64
            gender: m
            location: toscana
            education: scuola secondaria di II grado
            profession: lavoratore indipendente
          type: array
          items:
            $ref: '#/definitions/predict_needs_request'
      responses:
        '200':
          description: ''
          headers: {}
  /predict_attitudes:
    post:
      summary: predict_attitudes
      tags:
      - prediction
      operationId: predict_attitudes
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          example:
          - id: 1
            age: 64
            gender: m
            location: toscana
            education: scuola secondaria di II grado
            profession: lavoratore indipendente
          type: array
          items:
            $ref: '#/definitions/predict_needs_request'
      responses:
        '200':
          description: ''
          headers: {}
  /predict_status:
    post:
      summary: predict_status
      tags:
      - prediction
      operationId: predict_status
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          example:
          - id: 1
            age: 64
            gender: m
            location: toscana
            education: scuola secondaria di II grado
            profession: lavoratore indipendente
          type: array
          items:
            $ref: '#/definitions/predict_needs_request'
      responses:
        '200':
          description: ''
          headers: {}
  /read_samples:
    get:
      summary: read_samples
      tags:
      - samples and training
      operationId: read_samples
      deprecated: false
      produces:
      - application/json
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
  /read_metrics:
    get:
      summary: read_metrics
      tags:
      - samples and training
      operationId: read_metrics
      deprecated: false
      produces:
      - application/json
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
  /insert_samples:
    post:
      summary: Insert_samples
      tags:
      - samples and training
      operationId: Insert_samples
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          example:
          - sample_source: test
            sample_date: 02/01/2023
            sample:
            - needs:
                payment_financing_need: 0.9
                loan_financing_need: 0.4
                mortgage_financing_need: 0.7
              sociodemographics:
                age: 43
                gender: f
                location: piemonte
                education: scuola primaria
                profession: non occupato
            - attitudes:
                bank_activity_index: 0.2
                digital_activity_index: 0.9
              needs:
                payment_financing_need: 0.9
                loan_financing_need: 0.4
                mortgage_financing_need: 0.7
              sociodemographics:
                age: 49
                gender: m
                location: piemonte
                education: scuola primaria
                profession: non occupato
          type: array
          items:
            $ref: '#/definitions/Insert_samples_Request'
      responses:
        '200':
          description: ''
          headers: {}
  /train_models:
    put:
      summary: train_models
      tags:
      - samples and training
      operationId: train_models
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/train_models_request'
      responses:
        '200':
          description: ''
          headers: {}
  /delete_samples:
    delete:
      summary: delete_samples
      tags:
      - samples and training
      operationId: delete_samples
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/delete_samples_request'
      responses:
        '200':
          description: ''
          headers: {}
definitions:
  predict_needs_request:
    title: predict_needs_request
    example:
      id: 1
      age: 64
      gender: m
      location: toscana
      education: scuola secondaria di II grado
      profession: lavoratore indipendente
    type: object
    properties:
      id:
        type: integer
        format: int32
      age:
        type: integer
        format: int32
      gender:
        type: string
      location:
        type: string
      education:
        type: string
      profession:
        type: string
    required:
    - id
    - age
    - gender
    - location
    - education
    - profession
  Insert_samples_Request:
    title: Insert_samples_Request
    example:
      sample_source: test
      sample_date: 02/01/2023
      sample:
      - needs:
          payment_financing_need: 0.9
          loan_financing_need: 0.4
          mortgage_financing_need: 0.7
        sociodemographics:
          age: 43
          gender: f
          location: piemonte
          education: scuola primaria
          profession: non occupato
      - attitudes:
          bank_activity_index: 0.2
          digital_activity_index: 0.9
        needs:
          payment_financing_need: 0.9
          loan_financing_need: 0.4
          mortgage_financing_need: 0.7
        sociodemographics:
          age: 49
          gender: m
          location: piemonte
          education: scuola primaria
          profession: non occupato
    type: object
    properties:
      sample_source:
        type: string
      sample_date:
        type: string
      sample:
        type: array
        items:
          $ref: '#/definitions/Sample'
    required:
    - sample_source
    - sample_date
    - sample
  Sample:
    title: Sample
    example:
      needs:
        payment_financing_need: 0.9
        loan_financing_need: 0.4
        mortgage_financing_need: 0.7
      sociodemographics:
        age: 43
        gender: f
        location: piemonte
        education: scuola primaria
        profession: non occupato
    type: object
    properties:
      needs:
        $ref: '#/definitions/Needs'
      sociodemographics:
        $ref: '#/definitions/Sociodemographics'
      attitudes:
        $ref: '#/definitions/Attitudes'
    required:
    - needs
    - sociodemographics
  Needs:
    title: Needs
    example:
      payment_financing_need: 0.9
      loan_financing_need: 0.4
      mortgage_financing_need: 0.7
    type: object
    properties:
      payment_financing_need:
        type: number
        format: double
      loan_financing_need:
        type: number
        format: double
      mortgage_financing_need:
        type: number
        format: double
    required:
    - payment_financing_need
    - loan_financing_need
    - mortgage_financing_need
  Sociodemographics:
    title: Sociodemographics
    example:
      age: 43
      gender: f
      location: piemonte
      education: scuola primaria
      profession: non occupato
    type: object
    properties:
      age:
        type: integer
        format: int32
      gender:
        type: string
      location:
        type: string
      education:
        type: string
      profession:
        type: string
    required:
    - age
    - gender
    - location
    - education
    - profession
  Attitudes:
    title: Attitudes
    example:
      bank_activity_index: 0.2
      digital_activity_index: 0.9
    type: object
    properties:
      bank_activity_index:
        type: number
        format: double
      digital_activity_index:
        type: number
        format: double
    required:
    - bank_activity_index
    - digital_activity_index
  train_models_request:
    title: train_models_request
    example:
      predict_set:
        mean_age: 60
        mode_gender: f
        mode_location: lombardia
        mode_profession: non occupato
        mode_education: scuola secondaria di II grado
    type: object
    properties:
      predict_set:
        $ref: '#/definitions/PredictSet'
    required:
    - predict_set
  PredictSet:
    title: PredictSet
    example:
      mean_age: 60
      mode_gender: f
      mode_location: lombardia
      mode_profession: non occupato
      mode_education: scuola secondaria di II grado
    type: object
    properties:
      mean_age:
        type: integer
        format: int32
      mode_gender:
        type: string
      mode_location:
        type: string
      mode_profession:
        type: string
      mode_education:
        type: string
    required:
    - mean_age
    - mode_gender
    - mode_location
    - mode_profession
    - mode_education
  delete_samples_request:
    title: delete_samples_request
    example:
      ids:
      - 10
      - 11
      - 12
    type: object
    properties:
      ids:
        type: array
        items:
          type: integer
          format: int32
    required:
    - ids
tags:
- name: prediction
- name: samples and training